requirements.txt
aiohttp~=3.8.1
DateTime~=4.4
humanfriendly~=10.0
psycopg2~=2.9.3
requests~=2.27.1
python-dotenv~=0.20.0
py-cord~=2.1.3
